# Improve performance for `cudf::contains` when searching for a scalar