@ttnghia 2022-07-06
Here is the benchmark results:
```
Benchmark                                                              Time             CPU      Time Old      Time New       CPU Old       CPU New
---------------------------------------------------------------------------------------------------------------------------------------------------
Contains/SearchScalar_AllValid/32768/manual_time                    -0.4414         -0.3216             0             0             0             0
Contains/SearchScalar_AllValid/262144/manual_time                   -0.3719         -0.2666             0             0             0             0
Contains/SearchScalar_AllValid/2097152/manual_time                  -0.4689         -0.4047             0             0             0             0
Contains/SearchScalar_AllValid/16777216/manual_time                 -0.7573         -0.7391             1             0             1             0
Contains/SearchScalar_AllValid/134217728/manual_time                -0.8053         -0.8005             5             1             5             1
Contains/SearchScalar_AllValid/268435456/manual_time                -0.8057         -0.8054            10             2            10             2
Contains/SearchScalar_Nulls/32768/manual_time                       -0.3815         -0.2611             0             0             0             0
Contains/SearchScalar_Nulls/262144/manual_time                      -0.3703         -0.2673             0             0             0             0
Contains/SearchScalar_Nulls/2097152/manual_time                     -0.4685         -0.4092             0             0             0             0
Contains/SearchScalar_Nulls/16777216/manual_time                    -0.7554         -0.7358             1             0             1             0
Contains/SearchScalar_Nulls/134217728/manual_time                   -0.8056         -0.8008             5             1             5             1
Contains/SearchScalar_Nulls/268435456/manual_time                   -0.8099         -0.8044            10             2            10             2

```