@ttnghia 2022-07-06
Here is the original benchmark results:

### Before:
```
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
Contains/SearchScalar_AllValid/32768/manual_time          0.049 ms        0.068 ms        11579
Contains/SearchScalar_AllValid/262144/manual_time         0.049 ms        0.067 ms        12872
Contains/SearchScalar_AllValid/2097152/manual_time        0.100 ms        0.113 ms         6656
Contains/SearchScalar_AllValid/16777216/manual_time       0.631 ms        0.644 ms         1042
Contains/SearchScalar_AllValid/134217728/manual_time       4.90 ms         4.91 ms          144
Contains/SearchScalar_AllValid/268435456/manual_time       9.80 ms         9.81 ms           71
Contains/SearchScalar_Nulls/32768/manual_time             0.044 ms        0.062 ms        15785
Contains/SearchScalar_Nulls/262144/manual_time            0.049 ms        0.066 ms        13046
Contains/SearchScalar_Nulls/2097152/manual_time           0.101 ms        0.114 ms         6572
Contains/SearchScalar_Nulls/16777216/manual_time          0.637 ms        0.650 ms         1007
Contains/SearchScalar_Nulls/134217728/manual_time          4.91 ms         4.92 ms          134
Contains/SearchScalar_Nulls/268435456/manual_time          9.82 ms         9.83 ms           71
```

### After:
```
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
Contains/SearchScalar_AllValid/32768/manual_time          0.028 ms        0.046 ms        25542
Contains/SearchScalar_AllValid/262144/manual_time         0.031 ms        0.049 ms        22731
Contains/SearchScalar_AllValid/2097152/manual_time        0.053 ms        0.067 ms        12746
Contains/SearchScalar_AllValid/16777216/manual_time       0.153 ms        0.168 ms         4571
Contains/SearchScalar_AllValid/134217728/manual_time      0.954 ms        0.980 ms          732
Contains/SearchScalar_AllValid/268435456/manual_time       1.90 ms         1.91 ms          375
Contains/SearchScalar_Nulls/32768/manual_time             0.027 ms        0.046 ms        25510
Contains/SearchScalar_Nulls/262144/manual_time            0.031 ms        0.049 ms        22713
Contains/SearchScalar_Nulls/2097152/manual_time           0.054 ms        0.067 ms        12385
Contains/SearchScalar_Nulls/16777216/manual_time          0.156 ms        0.172 ms         4496
Contains/SearchScalar_Nulls/134217728/manual_time         0.955 ms        0.981 ms          733
Contains/SearchScalar_Nulls/268435456/manual_time          1.87 ms         1.92 ms          375
```